fully functional productsadmincontroller and views

+++ Gartenkraft/Areas/Admin/Controllers/AdminControllers/ProductsAdminController.cs
@@ -14,7 +14,7 @@ namespace Gartenkraft.Areas.Admin.Controllers.AdminControllers
     {
         private GartenkraftEntities db = new GartenkraftEntities();
 
-        // GET: Products
+        // GET: Admin/ProductsAdmin
         public ActionResult Index()
         {
             var tblProducts = db.tblProducts.Include(t => t.tblProduct_Line).Include(t => t.tblProduct_Category);
@@ -23,7 +23,7 @@ namespace Gartenkraft.Areas.Admin.Controllers.AdminControllers
             return View(tblProducts.ToList());
         }
 
-        // GET: Products/Details/5
+        // GET: Admin/ProductsAdmin/Details/5
         public ActionResult Details(int? id)
         {
             if (id == null)
@@ -39,7 +39,7 @@ namespace Gartenkraft.Areas.Admin.Controllers.AdminControllers
             return View(tblProduct);
         }
 
-        // GET: Products/Create
+        // GET: Admin/ProductsAdmin/Create
         public ActionResult Create()
         {
             ViewBag.product_line_id = new SelectList(db.tblProduct_Line, "product_line_id", "product_line_name");
@@ -47,7 +47,7 @@ namespace Gartenkraft.Areas.Admin.Controllers.AdminControllers
             return View();
         }
 
-        // POST: Products/Create
+        // POST: Admin/ProductsAdmin/Create
         // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
@@ -66,8 +66,8 @@ namespace Gartenkraft.Areas.Admin.Controllers.AdminControllers
             return View(tblProduct);
         }
 
-        // GET: Products/Edit/5
-        public ActionResult Edit(int? id, string message)
+        // GET: Admin/ProductsAdmin/Edit/5
+        public ActionResult Edit(int? id, string message, string errorMessage)
         {
             if (id == null)
             {
@@ -83,33 +83,47 @@ namespace Gartenkraft.Areas.Admin.Controllers.AdminControllers
             ViewBag.ProductImages = db.tblProduct_Image.Where(pi => pi.product_id == id).ToList();
             if (message != null)
             {
-                ViewBag.Message = message;
+                ViewBag.UploadErrorMessage = message;
+            }
+            if (errorMessage != null)
+            {
+                ViewBag.ErrorMessage = errorMessage;
             }
             tblProduct.SetPriceRange();
             return View(tblProduct);
         }
 
-        // POST: Products/Edit/5
+        // POST: Admin/ProductsAdmin/Edit/5
         // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "product_id,product_name,product_short_description,product_long_description,product_unit_cost,product_unit_price,product_category_id,product_line_id,product_weight,product_date_added,soft_delete,is_visible")] tblProduct tblProduct)
+        public ActionResult Edit([Bind(Include = "product_id,product_name,product_short_description,product_long_description,product_category_id,product_line_id,product_date_added,soft_delete,is_visible,is_custom_product")] tblProduct tblProduct)
         {
+            string errorMessage = "";
             if (ModelState.IsValid)
             {
-                db.Entry(tblProduct).State = EntityState.Modified;
-                db.SaveChanges();
-                return RedirectToAction("Index");
+                var prodOptions = db.tblProduct_Option.Where(o => o.product_id == tblProduct.product_id).ToList();
+                
+                // check if product type is valid
+                if (tblProduct.is_custom_product == false && prodOptions.Count > 1) { errorMessage = "This product is not qualified to be a simple product as it has multiple product options"; }
+
+                if (errorMessage == "")
+                {
+                    db.Entry(tblProduct).State = EntityState.Modified;
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
+                }
             }
             tblProduct.SetPriceRange();
+            ViewBag.ErrorMessage = errorMessage;
             ViewBag.product_line_id = new SelectList(db.tblProduct_Line, "product_line_id", "product_line_name", tblProduct.product_line_id);
             ViewBag.product_category_id = new SelectList(db.tblProduct_Category, "category_id", "category_name", tblProduct.product_category_id);
             ViewBag.ProductImages = db.tblProduct_Image.Where(pi => pi.product_id == tblProduct.product_id).ToList();
             return View(tblProduct);
         }
 
-        // GET: Products/Delete/5
+        // GET: Admin/ProductsAdmin/Delete/5
         public ActionResult Delete(int? id)
         {
             if (id == null)
@@ -124,32 +138,53 @@ namespace Gartenkraft.Areas.Admin.Controllers.AdminControllers
             return View(tblProduct);
         }
 
-        // POST: Products/Delete/5
+        // POST: Admin/ProductsAdmin/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
+            var invoicesContainsProduct = db.tblSales_Invoice_Lineitem.Where(li => li.product_id == id).ToList();
             tblProduct tblProduct = db.tblProducts.Find(id);
-            db.tblProducts.Remove(tblProduct);
-            db.SaveChanges();
+
+            // check if product had not been purchased before
+            if (invoicesContainsProduct == null || invoicesContainsProduct.Count == 0)
+            {
+                //remove product
+                db.tblProducts.Remove(tblProduct);
+                db.SaveChanges();
+            }
+            else
+            {
+                tblProduct.soft_delete = true;
+                db.Entry(tblProduct).State = EntityState.Modified;
+                db.SaveChanges();
+            }
             return RedirectToAction("Index");
         }
 
         #region Product options
 
-        // Get: Product/CreateOption
+        // Get: Admin/ProductsAdmin/CreateOption
         public ActionResult CreateOption(int? id)
         {
+            var product = db.tblProducts.Find(id);
+            var prodOptions = db.tblProduct_Option.Where(o => o.product_id == id).ToList();
+            if (product.is_custom_product == false && prodOptions.Count == 1)
+            {
+                string errorMessage = "This is a simple product. Please change to custom product before adding options.";
+                return RedirectToAction("Edit", new { id = id, errorMessage = errorMessage });
+            }
             var option = new tblProduct_Option() { product_id = Convert.ToInt32(id), tblProduct = db.tblProducts.Find(Convert.ToInt32(id)) };
             return View(option);
         }
 
-        // POST: Product/CreateOption
+        // POST: Admin/ProductsAdmin/CreateOption
         [HttpPost]
         [ValidateAntiForgeryToken]
         public ActionResult CreateOption([Bind(Include = "product_id,title, weight, unit_cost, unit_price")]tblProduct_Option option)
         {
             option.tblProduct = db.tblProducts.Find(option.product_id);
+
             if (ModelState.IsValid)
             {
                 db.tblProduct_Option.Add(option);
@@ -157,6 +192,65 @@ namespace Gartenkraft.Areas.Admin.Controllers.AdminControllers
                 return RedirectToAction("Edit", new { id = option.product_id });
             }
             return View(option);
+        }
+
+        // GET: Admin/ProductsAdmin/EditOption/5
+        public ActionResult EditOption(int? id)
+        {
+            if (id == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
+            tblProduct_Option tblProduct_Option = db.tblProduct_Option.Find(id);
+            if (tblProduct_Option == null)
+            {
+                return HttpNotFound();
+            }
+            ViewBag.product_id = new SelectList(db.tblProducts, "product_id", "product_name", tblProduct_Option.product_id);
+            return View(tblProduct_Option);
+        }
+
+        // POST: Admin/ProductsAdmin/EditOption/5
+        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
+        // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public ActionResult EditOption([Bind(Include = "title,weight,unit_cost,unit_price,product_id,option_id")] tblProduct_Option tblProduct_Option)
+        {
+            if (ModelState.IsValid)
+            {
+                db.Entry(tblProduct_Option).State = EntityState.Modified;
+                db.SaveChanges();
+                return RedirectToAction("Edit", new { id = tblProduct_Option.product_id });
+            }
+            ViewBag.product_id = new SelectList(db.tblProducts, "product_id", "product_name", tblProduct_Option.product_id);
+            return View(tblProduct_Option);
+        }
+
+        // GET: Admin/ProductsAdmin/DeleteOption/5
+        public ActionResult DeleteOption(int? id)
+        {
+            if (id == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
+            tblProduct_Option tblProduct_Option = db.tblProduct_Option.Find(id);
+            if (tblProduct_Option == null)
+            {
+                return HttpNotFound();
+            }
+            return View(tblProduct_Option);
+        }
+
+        // POST: Admin/ProductsAdmin/DeleteOption/5
+        [HttpPost, ActionName("DeleteOption")]
+        [ValidateAntiForgeryToken]
+        public ActionResult DeleteConfirmedOption(int id)
+        {
+            tblProduct_Option tblProduct_Option = db.tblProduct_Option.Find(id);
+            db.tblProduct_Option.Remove(tblProduct_Option);
+            db.SaveChanges();
+            return RedirectToAction("Edit", new { id = tblProduct_Option.product_id });
         }
 
         #endregion
